[Modified] OC-1225 #comment Changed backOffMs type to Long from Integer #time 5m

+++ src/backend/src/main/java/com/becon/opencelium/backend/resource/execution/RetryEx.java
@@ -3,7 +3,7 @@ package com.becon.opencelium.backend.resource.execution;
 public class RetryEx {
     private Integer maxAttempts;
 
-    private Integer backOffMs;
+    private Long backOffMs;
 
     public Integer getMaxAttempts() {
         return maxAttempts;
@@ -13,11 +13,11 @@ public class RetryEx {
         this.maxAttempts = maxAttempts;
     }
 
-    public Integer getBackOffMs() {
+    public Long getBackOffMs() {
         return backOffMs;
     }
 
-    public void setBackOffMs(Integer backOffMs) {
+    public void setBackOffMs(Long backOffMs) {
         this.backOffMs = backOffMs;
     }
 }
